requirements.txt
openai>=1.0.0         # For interacting with OpenAI compatible LLMs
ollama>=0.1.0         # For local LLM support
python-dotenv>=0.19.0 # For loading .env files
pypdf>=3.0.0          # For reading and manipulating PDF pages (PyPDF2 successor)
reportlab>=3.6.0      # For drawing text/stamps on PDFs
Pillow>=9.0.0         # reportlab dependency, good to have explicitly
chromadb>=0.4.0       # Vector database for semantic search
pdf2image>=1.16.0     # PDF to image conversion for vision OCR
PyMuPDF>=1.23.0       # Alternative PDF library for repair operations
pikepdf>=8.0.0        # Robust PDF library for repair and recovery
langchain>=0.1.0      # Core LangChain framework
langchain-community>=0.0.1  # Community integrations (legacy)
langchain-openai>=0.0.1  # OpenAI integration for ChatOpenAI
langchain-text-splitters>=0.0.1  # Advanced text chunking
langchain-ollama>=0.1.0   # New dedicated Ollama integration package
langchain-chroma>=0.1.0   # New dedicated Chroma integration package
pydantic>=2.0.0       # For structured output models
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Python
pytest>=7.0.0         # Testing framework